Split media details screen into components
